Moves main app to belong to django project
